requirements.txt
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
imutils==0.5.4
Pillow==9.5.0
face-recognition==1.3.0
dlib @ Dlib-main/dlib-19.22.99-cp37-cp37m-win_amd64.whl
